FFL-2664 Extract rules evaluator into core
