+++ packages/core/src/evaluation/errors.ts
@@ -1,0 +1,6 @@
+export class TargetingKeyMissingError extends Error {
+  constructor() {
+    super('Targeting key is required for split evaluation')
+    this.name = 'TargetingKeyMissingError'
+  }
+}
