+++ packages/node-server/src/configuration/evaluationMetadata.ts
@@ -1,6 +1,1 @@
-import type { PrecomputedFlagMetadata } from '@datadog/flagging-core'
-import type { TimeStamp } from '@datadog/js-core/time'
-
-export function createEvaluationTimestampMetadata(evaluationTimestampMs: TimeStamp): PrecomputedFlagMetadata {
-  return { __dd_eval_timestamp_ms: evaluationTimestampMs } as PrecomputedFlagMetadata
-}
+export { createEvaluationTimestampMetadata } from '@datadog/flagging-core'
